+++ package.json
@@ -1,39 +1,39 @@
 {
-	"name": "akore",
-	"version": "1.5.2",
-	"description": "The core of akita.ts",
-	"bin": {
-		"akore": "./dist/cli.js"
-	},
-	"main": "dist/index.js",
-	"scripts": {
-		"release": "npm version patch && npm publish"
-	},
-	"author": "Kodeko Studios",
-	"license": "MIT",
-	"devDependencies": {
-		"@types/hjson": "2.4.6",
-		"@types/js-yaml": "4.0.9",
-		"@types/lodash": "4.14.202",
-		"@types/node": "20.11.19",
-		"@types/uglify-js": "3.17.4",
-		"eslint": "9.0.0",
-		"prettier": "3.2.5",
-		"ts-node": "10.9.2",
-		"typescript": "5.4.2",
-		"webpack": "5.90.3",
-		"webpack-cli": "5.1.4"
-	},
-	"dependencies": {
-		"colors": "1.4.0",
-		"commander": "12.0.0",
-		"hjson": "3.2.2",
-		"js-yaml": "4.1.0",
-		"lodash": "4.17.21",
-		"module-alias": "2.2.3",
-		"ts-loader": "9.5.1",
-		"tsconfig-paths": "4.2.0",
-		"tslib": "2.6.2",
-		"uglify-js": "3.17.4"
-	}
+  "name": "akore",
+  "version": "1.5.2",
+  "description": "The core of akita.ts",
+  "bin": {
+    "akore": "./dist/cli.js"
+  },
+  "main": "dist/index.js",
+  "scripts": {
+    "release": "npm version patch && npm publish"
+  },
+  "author": "Kodeko Studios",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/hjson": "2.4.6",
+    "@types/js-yaml": "4.0.9",
+    "@types/lodash": "4.14.202",
+    "@types/node": "20.11.19",
+    "@types/uglify-js": "3.17.4",
+    "eslint": "9.0.0",
+    "prettier": "3.2.5",
+    "ts-node": "10.9.2",
+    "typescript": "5.4.2",
+    "webpack": "5.90.3",
+    "webpack-cli": "5.1.4"
+  },
+  "dependencies": {
+    "colors": "^1.4.0",
+    "commander": "12.0.0",
+    "hjson": "3.2.2",
+    "js-yaml": "4.1.0",
+    "lodash": "4.17.21",
+    "module-alias": "2.2.3",
+    "ts-loader": "9.5.1",
+    "tsconfig-paths": "4.2.0",
+    "tslib": "2.6.2",
+    "uglify-js": "3.17.4"
+  }
 }
