+++ src/index.ts
@@ -1,31 +1,31 @@
 import { register } from "tsconfig-paths";
 register();
 
-import $function from "./instructions/types/function";
-import $string from "./instructions/types/string";
-import $object from "./instructions/types/object";
 import $array from "./instructions/types/array";
+import $function from "./instructions/types/function";
 import $null from "./instructions/types/null";
+import $object from "./instructions/types/object";
+import $string from "./instructions/types/string";
 
-import $compound from "./instructions/variables/compound";
 import $call from "./instructions/variables/call";
+import $compound from "./instructions/variables/compound";
 import $get from "./instructions/variables/get";
 import $var from "./instructions/variables/var";
 
 import $escape from "./instructions/escape";
-import $print from "./instructions/print";
-import $new from "./instructions/new";
 import $if from "./instructions/if";
+import $new from "./instructions/new";
+import $print from "./instructions/print";
 
 import $modulo from "./instructions/util/modulo";
 import $multi from "./instructions/util/multi";
-import $sum from "./instructions/util/sum";
 import $sub from "./instructions/util/sub";
+import $sum from "./instructions/util/sum";
 
-import $continue from "./instructions/loops/continue";
-import $while from "./instructions/loops/while";
 import $break from "./instructions/loops/break";
+import $continue from "./instructions/loops/continue";
 import $for from "./instructions/loops/for";
+import $while from "./instructions/loops/while";
 
 export const BasicInstructions = {
 	$compound,
@@ -52,3 +52,4 @@ export const BasicInstructions = {
 };
 
 export * from "./classes/";
+export { transpileFiles } from "./helpers/transpile_files";
